+++ Demo/Client/_Imports.razor
@@ -11,7 +11,7 @@
 @using Demo.Client.Shared
 @using Demo.Client.Services
 @using Demo.Shared
-@using Demo.Shared.Requests
+@using Demo.Shared.Playground
 
 @using Pipaslot.Mediator
