+++ Demo/Shared/Playground/MessageWithNotification.cs
@@ -1,0 +1,9 @@
+﻿using Pipaslot.Mediator;
+
+namespace Demo.Shared.Playground
+{
+    public class MessageWithNotification : IMessage
+    {
+        public bool Fail { get; set; }
+    }
+}
