--- Demo/Shared/Requests/DemoMessage.cs
@@ -1,9 +1,0 @@
-﻿using Pipaslot.Mediator;
-
-namespace Demo.Shared.Requests
-{
-    public class DemoMessage : IMessage
-    {
-        public bool Fail { get; set; }
-    }
-}
